\docType{package}
\name{growth}
\alias{growth}
\alias{growth-package}
\title{growth: Nonlinear Growth Models}
\description{
  A recopilation of nonlinear growth models used in many
  branches of science
}
\details{
  \tabular{ll}{ Package: \tab growth \cr Version: \tab
  0.0.1 \cr License: \tab GPL-3 \cr }
}
\author{
  Daniel Rodriguez Perez
  \email{daniel.rodriguez.perez@gmail.com}
}

