% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthmodels.R
\docType{package}
\name{growthmodels-package}
\alias{growthmodels-package}
\alias{growthmodels}
\title{growthmodels: Nonlinear Growth Models}
\description{
A compilation of nonlinear growth models.
}
\details{
\tabular{ll}{
  Package: \tab growthmodels \cr
  Version: \tab 1.2.0 \cr
  License: \tab GPL-3 \cr
}
}
\references{
D. Fekedulegn, M. Mac Siurtain, and J. Colbert, "Parameter estimation of
nonlinear growth models in forestry," Silva Fennica, vol. 33, no. 4, pp.
327-336, 1999.

M. M. Kaps, W. O. W. Herring, and W. R. W. Lamberson, "Genetic and
environmental parameters for traits derived from the Brody growth curve and
their relationships with weaning weight in Angus cattle.," Journal of
Animal Science, vol. 78, no. 6, pp. 1436-1442, May 2000.

A. Tsoularis and J. Wallace, "Analysis of logistic growth models.,"
Math Biosci, vol. 179, no. 1, pp. 21-55, Jul. 2002.

A. Khamiz, Z. Ismail, and A. T. Muhammad, "Nonlinear growth models for
modeling oil palm yield growth," Journal of Mathematics and Statistics,
vol. 1, no. 3, p. 225, 2005.

Michael J. Panik, "Growth Curve Modeling: Theory and Applications",
John Wiley & Sons, December 2013.

http://en.wikipedia.org/wiki/Generalised_logistic_function
}
\author{
Daniel Rodriguez \email{daniel.rodriguez.perez@gmail.com}
}
