% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosso.R
\name{rosso}
\alias{rosso}
\alias{rosso.inverse}
\title{Rosso growth model}
\usage{
rosso(t, mu, lag, A, U)

rosso.inverse(x, mu, lag, A, U)
}
\arguments{
\item{t}{time}

\item{mu}{maximal growth rate}

\item{lag}{time lag}

\item{A}{the lower asymptote}

\item{U}{the upper asymptote}

\item{x}{size}
}
\description{
Computes the Rosso growth model and its inverse
\deqn{ y(t) = U - \log(1 + (\exp(U - A) - 1)*\exp(-mu*(t-lag))) }
}
\examples{
growth <- rosso(0:10, 0.05, 2, 1, 10)

time <- rosso.inverse(growth, 0.05, 2, 1, 10)

}
\references{
T. Ross, "Indices for performance evaluation of predictive models in food
microbiology." J. Appl. Bacteriol. vol. 81, no. 5, pp. 501-508. Nov. 1996.
}
\author{
Daniel Rodriguez
}
