\docType{package}
\name{growthmodels-package}
\alias{growthmodels}
\alias{growthmodels-package}
\title{growthmodels: Nonlinear Growth Models}
\description{
  A recopilation of nonlinear growth models used in many
  areas
}
\details{
  \tabular{ll}{ Package: \tab growthmodels \cr Version:
  \tab 1.0.0 \cr License: \tab GPL-3 \cr }
}
\author{
  Daniel Rodriguez Perez
  \email{daniel.rodriguez.perez@gmail.com}
}
\references{
  D. Fekedulegn, M. Mac Siurtain, and J. Colbert,
  "Parameter estimation of nonlinear growth models in
  forestry," Silva Fennica, vol. 33, no. 4, pp. 327-336,
  1999.

  http://en.wikipedia.org/wiki/Generalised_logistic_function
}

